requirements.txt
transformers
torch
Pillow
pypianoroll
tqdm
pretty_midi